backend/hanachan/requirements.txt
# Hanachan Backend Requirements
flask>=2.0.0
flask-cors>=3.0.0
flask-sqlalchemy>=3.0.0
flask-talisman>=1.0.0
flask-limiter>=3.0.0
pydantic>=2.0.0
pymongo>=4.0.0
psycopg2-binary>=2.9.0
pyjwt
pyyaml
python-dotenv
requests
PyPDF2
pytest==7.4.0
pytest-mock==3.11.1
langchain
langchain-community
langchain-ollama
langchain-neo4j
qdrant-client
langchain-qdrant
langgraph
rich
langchain-openai
redis
rq
python-magic
